Implements error boundary and 404 page
Introduces an error boundary component to gracefully handle unexpected errors and prevent the entire application from crashing.
Provides a default error fallback UI, with the option to provide a custom fallback.

Adds a 404 "Not Found" page to handle invalid routes.

Updates the router configuration to include the error boundary and 404 page.

Fixes issues with data fetching in customer and order services and store.
The API response objects are returned instead of response.data.